🪦 feat(api/common/constants): remove DEFAULT_USER_GENDER constant and refactor related imports
