--- apps/api/src/common/constants/default.ts
@@ -1,1 +1,0 @@
-export const DEFAULT_USER_GENDER = "man";
