+++ apps/api/src/domain/entities/session.ts
@@ -1,5 +1,11 @@
-import { sessionExpiresSpan, sessionRefreshSpan } from "../../common/constants";
+import { TimeSpan } from "../../common/utils";
 import type { SessionId, UserId } from "../value-object";
+
+export const SESSION_EXPIRES_SPAN_DAYS = 30 as const;
+export const SESSION_REFRESH_SPAN_DAYS = 15 as const;
+
+export const sessionExpiresSpan = new TimeSpan(SESSION_EXPIRES_SPAN_DAYS, "d");
+export const sessionRefreshSpan = new TimeSpan(SESSION_REFRESH_SPAN_DAYS, "d");
 
 export interface Session {
 	id: SessionId;
